Give up on trying to keep the range intact.
It would need more testing to know when it can be relied upon. Also, in
complex cases, e.g. a highlight in a highlight, it may fail and throw.
We could simply catch and quench the errors, but a range that *might*
still be pointing at the same thing it did before seems of little use.

Could figure this out and improve this another time.